fix: entry ID for non-ldap persistence
